scraper_service/requirements.txt
pip==22.3.1
setuptools==65.5.1
wheel==0.38.4
Flask==3.0.0
PyMongo
PyGithub

